+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 For more information about changelogs, check [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 1.4.3 - 2026-06-24
+
+* [Feature] Add "Brands" icons
+
 ## 1.4.2 - 2026-06-24
 
 * [Feature] Add common recourse icons